regrouped and improved placement of keys
When entering rhythms, shift plus 1, 2 and 4 will enter these corresponding notes, Right shift  plus 3 is reserved for the dot. This leaves number 5 to be an 8th note and number 6 to be a 16th note which makes typing and rhythm entry more intuitive.
Also, added right shift plus numpad enter which finishes a chord by pressing the end key and then adding a space.
Updated the help contained in the program so it reflects these changes.